requirements.txt
FlagEmbedding
peft
flask
tqdm
pymysql
meilisearch
canal-python
mysql-connector-python